+++ codebase/boilercontroller_components/boilercontroller_module.py
@@ -1,0 +1,7 @@
+from . import boilercontroller_class as BoilerControllerClass
+
+
+def createboilercontroller():
+	
+	return BoilerControllerClass.DefineBoilerController()
+
