+++ codebase/boilercontroller_components/boilerswitch_component/boilerswitch_class.py
