+++ codebase/boilercontroller_components/thermometer_component/thermometer_module.py
