--- codebase/display_component/board_subcomponent/board_module.py
@@ -1,5 +1,0 @@
-import board_class as BoardClass
-
-def createboard():
-	return BoardClass.DefineBoard()
-
