--- codebase/display_component/buttons_subcomponent/buttons_module.py
@@ -1,5 +1,0 @@
-import buttons_class as ButtonsClass
-
-def createbuttons(controls):
-	return ButtonsClass.DefineButtons(controls)
-
