--- codebase/display_component/display_module.py
@@ -1,5 +1,0 @@
-from . import display_class as DisplayClass
-
-def createdisplay(controls):
-	return DisplayClass.DefineDisplay(controls)
-
