--- codebase/display_component/runway_subcomponent/runway_module.py
@@ -1,5 +1,0 @@
-import runway_class as RunwayClass
-
-def createrunway():
-	return RunwayClass.DefineRunway()
-
