+++ codebase/miscellaneous_components/meteo_component/meteo_class.py
