+++ codebase/userinterface_components/controls_component/tempselector_subcomponent/tempselector_module.py
