+++ codebase/userinterface_components/displayboard_component/displayboard_module.py
@@ -1,0 +1,5 @@
+import displayboard_derivedclass as BoardClass
+
+def createboard():
+	return BoardClass.DefineBoard()
+
