+++ codebase/userinterface_components/displayboard_component/flame_subcomponent/flame_module.py
@@ -1,0 +1,5 @@
+import flame_class as FlameClass
+
+def createflame():
+	return FlameClass.DefineFlame()
+
