+++ codebase/userinterface_components/displaybuttons_component/displaybuttons_module.py
@@ -1,0 +1,5 @@
+import displaybuttons_derivedclass as ButtonsClass
+
+def createbuttons(controls):
+	return ButtonsClass.DefineButtons(controls)
+
