+++ codebase/userinterface_components/displayrunway_component/displayrunway_module.py
@@ -1,0 +1,5 @@
+import displayrunway_derivedclass as RunwayClass
+
+def createrunway():
+	return RunwayClass.DefineRunway()
+
