+++ codebase/userinterface_components/maindisplay_component/graphicdata_subcomponents/graphicdata_colours.py
