+++ codebase/userinterface_components/maindisplay_component/maindisplay_module.py
@@ -1,0 +1,5 @@
+from . import maindisplay_class as DisplayClass
+
+def createdisplay(controls):
+	return DisplayClass.DefineDisplay(controls)
+
