+++ codebase/userinterface_components/userinterface_module.py
@@ -1,0 +1,7 @@
+from . import userinterface_class as UserInterfaceClass
+
+
+def createuserinterface():
+	
+	return UserInterfaceClass.DefineUserInterface()
+
